#3 #2 adds address infos and edits customer edit transactions

+++ skn_curtain_WebUI/skn_curtain_Data/Entities/Columns.cs
@@ -5,6 +5,6 @@
         public string ColumnWidth { get; set; }
         public string ColumnName { get; set; }
         public bool Status { get; set; }
-        public int CurtainInfoesId { get; set; }
+        public long CurtainInfoesId { get; set; }
     }
 }
